added aliases, project restuctured
