+++ dev/webpack.dev.config.js
@@ -4,7 +4,16 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 
 module.exports = {
-  entry: path.resolve(__dirname, '../src/renderer/index.js'),
+  entry: path.resolve(__dirname, '../src/renderer/app.js'),
+  resolve: {
+    alias: {
+      'Components': path.resolve(__dirname, '../src/renderer/components'),
+      'Pages': path.resolve(__dirname, '../src/renderer/pages'),  
+      'Actions': path.resolve(__dirname, '../src/renderer/actions'),  
+      'Reducers': path.resolve(__dirname, '../src/renderer/reducers'),
+      'Selectors': path.resolve(__dirname, '../src/renderer/selectors')
+    }
+  },
   mode: 'development',
   module: {
     rules: [
