--- src/renderer/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>  
-  <head>
-    <title>My App</title>
-    <!-- <link rel="stylesheet" type="text/css" href="../../node_modules/typeface-roboto/index.css"> -->
-  </head>
-  <body>
-    <div id="app"></div>
-    <script src="./index.js" data-container="#app"></script>
-    <script>
-      if (process.env.BROWSER_SYNC_CLIENT_URL) {
-        const current = document.currentScript;
-        const script = document.createElement('script');
-        script.src = process.env.BROWSER_SYNC_CLIENT_URL;
-        script.async = true;
-        current.parentNode.insertBefore(script, current);
-      }
-    </script>
-  </body>
-</html>
