+++ src/renderer/pages/HelpPage/HelpPageView.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import PageBase from 'Pages/common/PageBase';
+
+const HelpPageView = () => {
+  return (
+    <PageBase>I am help page</PageBase>
+  );
+};
+
+export default HelpPageView;
