+++ src/renderer/pages/JsonPage/JsonPageView.js
@@ -1,18 +1,14 @@
-const PageBase = require('../common/PageBase');
-
 import React from 'react';
-// import PageBase from 'Pages/common/PageBase';
+import PageBase from 'Pages/common/PageBase';
 import { Button } from '@material-ui/core';
 
 const JsonPage = ({ handleAddWatcher }) => {
-  console.log('render...');
-
   return (
-    // <PageBase>
+    <PageBase>
       <Button variant='contained' color='primary' onClick={handleAddWatcher}>
         Add watcher
       </Button>
-    // </PageBase>
+    </PageBase>
   );
 };
