+++ src/renderer/pages/JsonPage/index.js
@@ -1,23 +1,9 @@
-import path from 'path';
-import { remote } from 'electron';
-import { connect } from 'react-redux';
-import JsonPageComponent from './JsonPageComponent';
-const { BrowserWindow } = remote;
+import JsonPageView from './JsonPageView';
+import JsonPageRedux from './JsonPageRedux';
 
-const mapDispatchToProps = (dispatch) => ({
-  handleAddWatcher: () => {
-    const fileToRender = path.resolve(path.join(__dirname, '../renderer/index.html'));
-    const win = new BrowserWindow({ 
-      width: 400,
-      height: 200,
-      show: false
-    });
+export {
+  JsonPageView,
+  JsonPageRedux
+};
 
-    // win.on('close', () => win = null);
-    // win.loadFile(fileToRender);
-
-  }
-});
-
-
-export default connect(null, mapDispatchToProps)(JsonPageComponent);
+export default JsonPageRedux;
