--- src/renderer/pages/JsonPage/useStyles.js
@@ -1,8 +1,0 @@
-import { makeStyles } from "@material-ui/styles";
-
-const useStyles = makeStyles(theme => ({
-    button: {
-    }
-}));
-
-export default useStyles;
