+++ src/renderer/pages/common/PageBase/PageBaseStyles.js
@@ -1,9 +1,9 @@
 import { makeStyles } from "@material-ui/styles";
 
-const useStyles = makeStyles(theme => ({
+const PageBaseStyles = makeStyles(theme => ({
     container: {
         margin: '1.25rem'
     }
 }));
 
-export default useStyles;
+export default PageBaseStyles;
